TODO calculate implied price of an option
The implied volatility concept is dubious. Is the market value of an option reflecting the expected volatility through expiry?

One could calculate instead the implied expiry date or the implied dividend rate, etc.

But one estimate that may be sensible is the market's estimate for the closing price of the underlying at expiry.